requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
cachetools==6.2.1
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.13
constantly==23.10.4
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
daphne==4.0.0
decorator==4.4.2
Django==5.1.2
django-allauth==65.0.2
django-widget-tweaks==1.5.0
elevenlabs==1.0.0
exceptiongroup==1.3.0
executing==2.2.1
facenet-pytorch==2.5.3
fer==22.5.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.185.0
google-auth==2.42.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.62.2
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
incremental==24.7.2
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
keras==3.12.0
kiwisolver==1.4.9
kombu==5.5.4
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf!=4.21.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
pytz==2025.2
pywin32==311
redis==5.0.1
requests==2.32.5
rich==14.2.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
tomli==2.3.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.0.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
yarl==1.22.0
zope.interface==8.0.1